cvefeed: drop support for XML feeds
This commit drops the support for XML file format for nvd feeds.

It also unifies the JSON schema used in different pkgs to a single one,
distributed under cvefeed/jsonschema. This json schema is kept exactly
as it is generated (the previous one, used in nvdjson was modified to
satisfy some of the Interfaces).

The -feed flag was kept for backwards compatibility.