[cvss3] Better API for cvss3 vector
calling v1.parse(str) could fail in the middle so some of the values
would change and some wouldn't. There would be no way to recover.

This way we separate this into first parsing the vectors on each own and
then doing merge (absorb or absorbDefined)
Absorb will absorb all metrics from the new vector (even the undefined
ones). This is used for merging vectors from two sources when we think
that if the new source says the metric is undefined, that's the way we
want it.
AbsorbDefined will absorb only the values in the new vector which are
actually defined and override them in the first vector.