[cpe2cve] Optimization
One of the earlier pull requests introduced a performance regression.
The problem was that on each input cpe it was parsing all nodes in the
dictionary to construct wfn.Attributes. This is obviously a wrong
approach since we can parse them in the start and then just match on
input cpes.

This is done in the interfaces.go file using helper structs cpeMatch,
node and cveItem which are just wrappers around corresponding fields in
the NVD jsonchema.

Another "bug" was noticed, and fixed.
When the node had children, it wasn't parsed recursively so some
matching might not work as specified in the feed. Now it's fixed
(recursively calling `newNode()`)

Test: `go build && go test`
Also, tested on our production dataset. Before this commit, it took more
than 4 minutes to process 10 (huge) lines of input data.
After this commit, it takes 11 seconds.