[redhat] Implement a binary to filter out fixed packages based on the RedHat feed
RedHat feed tells us which rpm packages have been fixed for which
release for which CVE.
This binary implements this kind of filtering.
Input is in csv format and you specify which fields contain packages,
distribution and cve. Output is the same format as input, just has
packges filtered out. Only the ones that haven't been reported as fixed
stay in the output.

go test ./...