client: Always return a valid request id even if tracing is disabled
The http failure injection code uses the request number to decide if it should
fail a request or not. Request numbers start at 1. 0 is never assigned as a
request number and 0 is used to mean "no fault injection".

Unfortunatly, traceRequestStart was returning 0 when tracing was disabled,
which meant that, under normal usage, all requests were triggering the fault
injection code. Argh!