[redhat] Err on the side of caution when facing unknown resolution
If we ever encounter an unknown CVE status string, we were answering that the
problem was fixed. Better err on the side of caution and flag the CVE as not
fixed.

We don't want to have a hard parsing error that would stop the whole process,
just surface a warning to be picked up an fixed.