[redhat] Add the ability to list non applicable CVEs for a given package (#133)
* [rpm] Unexport VersionCompare

The linter was complaining this function wasn't documented. We can actually
unexport it as it has no user outside of the package.

* [rpm] Fix various small linter issues

* [redhat] Add a few doc strings

* [redhat] Fix instead typo

* [redhat] Internalize the CVE map in the Feed

I'd like to store some more (cached) state there and having the Feed map
immutable helps ensuring the cached state remains valid. A struct is also handy
to store the cached state along with the original feed data.

* [redhat] Add the ability to list fixed CVEs for a given (distro, package)

* [redhat] Add a redhat_query utility