[cvfeed] Provide utility to compute the CVE cache hit ratio
Currently, computing the key indexing the CVE cache depends on the order of the
list of CPEs. The cache hit ratio is a good way to evaluate how effective we
are at using the cache.

Sorting the incoming list of CPEs bumps the hit ration from ~68% to ~95% on a
random sample of 1000 list of CPEs and improves wall runtime by 10%.