[cvefeed] Sort CPEs when computing the cache key
We've established that sorting the CPEs lead to more cache hits (68% -> 95%).
Instead of having to do it in the client code and have that constraint in the
API, we can sort the CPEs when computing the cache key!

With this and a test 1000 unsorted lists of CPEs, we stay at 95% hit rate.