[cvss] Add functions to compare two CVSS vectors
To compare for equility, we want to do a bit more than just comparing the
string representations of vectors. For some of the metrics, not specifying a
score is the same as specifying one of the other possible values.

For instance Remediation Level (RL):

> Assigning this value indicates there is insufficient information to choose
> one of the other values, and has no impact on the overall Temporal Score,
> i.e., it has the same effect on scoring as assigning Unavailable.

So not specifying RL or setting RL:X is the same as setting RL:U.