[vulndb] Checking for errors after scanning all rows
From https://pkg.go.dev/database/sql#Rows.Next:
```
Next prepares the next result row for reading with the Scan method. It returns true on success, or false if there is no next result row or an error happened while preparing it. Err should be consulted to distinguish between the two cases.
```

Due to not checking this, a user might hit timeout issues and believe that they had the complete resultset. This diffs returns an error in this case