Fix #205.
Quoting original issue:
```
When version parts are compared, 0's are padded to the left to ensure
they have the same length. But there is a corner case of two versions
being semantically the same, but one carries zeros to the left, example:

    2023.02.13
    2023.2.13

When all parts match, the code fallback to string length to make a decision, which incorrectly chooses the longest string as latest.
```